Rename routerMock and implement all exports.
